% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_link.R
\name{data_link}
\alias{data_link}
\title{Run probabilistic linkage between SINAN and SIM}
\usage{
data_link(
  df_sinan,
  df_sim,
  dedup_sim = FALSE,
  filter_sogi = c("all", "lgbt", "gay", "lesbian", "bissexual", "travesti", "mul.trans",
    "hom.trans"),
  threshold = 0.35,
  save_summary = TRUE,
  summary_path = "linkage_summary.txt"
)
}
\arguments{
\item{df_sinan}{Data.frame or list of data.frames (SINAN)}

\item{df_sim}{Data.frame or list of data.frames (SIM)}

\item{dedup_sim}{Logical. Whether to deduplicate SIM}

\item{filter_sogi}{Filter for SOGI groups}

\item{threshold}{Matching threshold (default = 0.35)}

\item{save_summary}{Logical. Save linkage summary}

\item{summary_path}{File path for summary}
}
\value{
A data_link_pipe object
}
\description{
Run probabilistic linkage between SINAN and SIM
}
