% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_link.R
\name{data_link}
\alias{data_link}
\title{Realiza vinculação (linkage) entre bases de dados}
\usage{
data_link(
  df_sinan,
  df_sim,
  dedup_sim = FALSE,
  filter_sogi = c("all", "lgbt", "gay", "lesbian", "bissexual", "travesti", "mul.trans",
    "hom.trans"),
  threshold = 0.35,
  save_summary = TRUE,
  summary_path = "linkage_summary.txt"
)
}
\arguments{
\item{df_sinan}{Data frame contendo dados do SINAN}

\item{df_sim}{Data frame contendo dados do SIM}

\item{threshold}{Valor numérico para critério de pareamento}
}
\value{
Um data frame com os registros vinculados
}
\description{
Função para integrar dados provenientes de diferentes sistemas
de informação em saúde, como SINAN e SIM.
}
